+++ Makefile
@@ -6,13 +6,13 @@
 #    By: x230 <x230@student.42.fr>                  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/05/19 12:15:10 by fvonsovs          #+#    #+#              #
-#    Updated: 2023/05/30 16:21:54 by x230             ###   ########.fr        #
+#    Updated: 2023/05/30 17:06:24 by x230             ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 CFLAGS		=	-Wall -Wextra -Werror -fsanitize=address -g -L/usr/include -lreadline
 NAME		=	minishell
-SRCS		=	main.c minishell.c parser_lines.c utils_1.c test_functions.c
+SRCS		=	main.c minishell.c parser_lines.c utils_1.c utils_2.c test_functions.c
 OBJS		=	${SRCS:.c=.o}
 LIBC		=	ar -cq
 RM			=	rm -f
