+++ test_functions.c
@@ -6,20 +6,25 @@
 /*   By: x230 <x230@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/30 15:59:43 by x230              #+#    #+#             */
-/*   Updated: 2023/05/30 16:34:39 by x230             ###   ########.fr       */
+/*   Updated: 2023/05/30 17:04:20 by x230             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
 // dont forget to free after
-void test_parse_line(char *line)
+void test_parse_line(char **tokens)
 {
-    char **tokens = split_line(line);
-
     printf("Tokens found:\n");
     for (int i = 0; tokens[i] != NULL; i++)
     {
-        printf("%s\n", tokens[i]);
+        if (tokens[i] != NULL)
+        {
+            printf("%s\n", tokens[i]);
+        }
+        else
+        {
+            printf("Token %d is NULL\n", i);
+        }
     }
 }
